issue #73 : Data Converter changed to generate users, 'add-wildfly-users' script created

+++ data/src/main/java/OptionParser.java
@@ -52,9 +52,9 @@ public class OptionParser {
             }
 
             if (cmd.hasOption("u")) {
-                Long numberOfUsers = null;
+                Integer numberOfUsers = null;
                 try {
-                    numberOfUsers = Long.parseLong(cmd.getOptionValue("u"));
+                    numberOfUsers = Integer.parseInt(cmd.getOptionValue("u"));
                 } catch (NumberFormatException e) {
                     log.error("Failed to parse option as number");
                     help();
